pyproject.toml
[tool.poetry]
name = "pdf-to-md"
version = "0.1.0"
description = ""
authors = ["Vik Paruchuri <vik.paruchuri@gmail.com>"]
readme = "README.md"
packages = [{include = "pdf_to_md"}]

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.33.2"
scikit-learn = "^1.3.1"
ftfy = "^6.1.1"
PyMuPDF = "^1.23.3"
pydantic = "^2.3.0"
python-dotenv = "^1.0.0"
torch = "^2.0.1"
pydantic-settings = "^2.0.3"
ctransformers = "^0.2.27"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
